requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
torch
torchvision
numpy
pillow
scipy
opencv-python-headless
